requirements.txt
flask==3.0.2
torch==2.2.2
torchvision==0.17.2
pillow==10.3.0
numpy==1.26.4
joblib==1.4.2
